[step4] GameResult 파일 분리
